+++ core/modules/modDeliveryAddress.class.php
@@ -93,6 +93,7 @@ class modDeliveryAddress extends DolibarrModules
 		//							'workflow' => array('WORKFLOW_MODULE1_YOURACTIONTYPE_MODULE2'=>array('enabled'=>'! empty($conf->module1->enabled) && ! empty($conf->module2->enabled)', 'picto'=>'yourpicto@deliveryaddress')) // Set here all workflow context managed by module
 		//                        );
 		$this->module_parts = array(
+			'substitutions' => 1,
 			'hooks' => array(
 				'propalcard',
 				'ordercard',